+++ Changes.md
@@ -1,6 +1,13 @@
 1.6.x.x (relative to 1.6.14.1)
 =======
 
+Fixes
+-----
+
+- SceneInspector, AttributeEditor :
+  - Fixed bug causing history inspection to incorrectly include AttributeTweaks `CreateIfMissing` tweaks that have not modified the scene due to the attribute already existing upstream of the tweak.
+  - Fixed bug causing history inspection to incorrectly include Attributes nodes with an enabled `global` plug when inspecting attributes on scene locations.
+
 1.6.14.1 (relative to 1.6.14.0)
 ========